refactor knight controller
